Update DBConverter, Editor, and PolygonDrawer
This commit updates the DBConverter, Editor, and PolygonDrawer files.

In DBConverter, a new property "overlappingTiles" has been added to the Polygon type.

In Editor, the default value for "tileMode" has been changed to "Straight".

In PolygonDrawer, the calculation for "pieces" has been updated to use the area of a tile in square meters. The text display for the polygon area and pieces has been modified to include "OV" and "MV" values.

In HelpModal, a new section explaining the "OV Area" and "MV Area" calculations has been added.